requirement.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
djangorestframework-simplejwt
azure-identity
openai
python-dotenv
azure-ai-projects

